Allow PieMenuLayout to display a third value
before 0d78ef32d094cc63bff5299c85c7ebfc16c9dd51, only two values would be
added to a PieMenuLayout at a time, but since then it was possible for a
third value to be added to the PieMenu. So, this commit allows the PieMenu
to show that third value.