Rename 'move to current space' command
I feel like, but have not yet verified, that the right-click-deselct
is not discoverable. By happenstance, when a user has a token selected
and clicks the space on that token, the resulting menu option is a
no-op labeled "Move to this space". This commit makes the happenstance
explicit, and makes the name of the no-op "Cancel"